requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-socks==0.8.0
aiosignal==1.3.1
altgraph==0.17.3
amqp==5.1.1
aniso8601==9.0.1
appdirs==1.4.4
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.6.2
bs4==0.0.1
cchardet==2.1.7
celery==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
chromedriver-binary==114.0.5735.16.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coala-utils==0.5.1
colorama==0.4.6
comm==0.1.3
constantly==15.1.0
contourpy==1.1.1
cryptography==40.0.2
cssselect==1.2.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
EasyProcess==1.1
elastic-transport==8.4.0
elasticsearch==8.8.0
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fake-useragent==1.1.3
fastjsonschema==2.17.1
filelock==3.12.0
flake8==6.0.0
Flask==2.3.2
Flask-RESTful==0.3.10
fonttools==4.43.1
frozenlist==1.3.3
future==0.18.3
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.3.0
GetOldTweets3==0.0.11
googletrans==4.0.0rc1
googletransx==2.4.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==22.10.0
inexactsearch==1.0.2
ipykernel==6.24.0
ipython==8.14.0
isort==5.13.2
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.5
kombu==5.3.1
langdetect==1.0.9
Levenshtein==0.25.1
libretranslatepy==2.1.1
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
mypy==1.1.1
mypy-extensions==1.0.0
nbformat==5.9.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parse==1.19.0
parsel==1.8.1
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
plotly==5.15.0
prompt-toolkit==3.0.38
Protego==0.2.1
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycares==4.3.0
pycodestyle==2.10.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==8.2.2
pyflakes==3.0.1
pygame==2.3.0
PyGetWindow==0.0.9
Pygments==2.15.1
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pymongo==4.3.3
PyMsgBox==1.0.9
pyOpenSSL==23.1.1
pyparsing==3.1.1
PyPDF2==3.0.1
pyperclip==1.8.2
pyppeteer==1.0.2
pyquery==2.0.0
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.25.1
python-socks==2.3.0
pytweening==1.0.7
pytz==2023.3
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==25.1.0
queuelib==1.6.2
rapidfuzz==3.9.7
regex==2023.6.3
requests==2.29.0
requests-file==1.5.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
schedule==1.2.0
scikit-learn==1.2.2
scipy==1.10.1
Scrapy==2.9.0
selenium==4.24.0
service-identity==21.1.0
silpa-common==0.3
six==1.16.0
sniffio==1.3.0
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soundex==1.1.3
soupsieve==2.4.1
spellchecker==0.4
stack-data==0.6.2
sympy==1.13.1
tenacity==8.2.2
threadpoolctl==3.1.0
tldextract==3.4.4
toml==0.10.2
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
translate==3.6.1
trimesh==4.1.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.14.0
tweetscraper==1.2.6
twint==2.1.20
Twisted==22.10.0
twisted-iocpsupport==1.0.3
twitter==1.19.6
twitterscraper==1.6.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
w3lib==2.1.1
wcwidth==0.2.6
webdriver-manager==3.8.6
websocket-client==1.8.0
websockets==10.4
Werkzeug==2.3.4
wsproto==1.2.0
yarl==1.9.2
yml==0.0.1
zipp==3.15.0
zope.interface==6.0
